editoast: isolate rolling stock import CLI into its own module
